requirements.txt
click==8.0.4
et-xmlfile==1.1.0
Flask==2.0.3
itsdangerous==2.1.2
Jinja2==3.1.0
MarkupSafe==2.1.1
numpy==1.22.3
openpyxl==3.0.9
pandas==1.4.2
protobuf==3.17.3
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
Werkzeug==2.0.3
